CTAA-1686 - Refactor of the cleanup method

+++ app/src/main/java/at/roteskreuz/stopcorona/model/managers/DatabaseCleanupManager.kt
@@ -10,8 +10,6 @@ import at.roteskreuz.stopcorona.utils.toRollingStartIntervalNumber
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import org.koin.standalone.KoinComponent
-import org.threeten.bp.Instant
-import org.threeten.bp.ZoneId
 import org.threeten.bp.ZonedDateTime
 import timber.log.Timber
 import kotlin.coroutines.CoroutineContext
@@ -29,8 +27,6 @@ class DatabaseCleanupManagerImpl(
 ) : DatabaseCleanupManager, CoroutineScope, KoinComponent {
 
     companion object {
-        private val UNIX_TIME_START: ZonedDateTime =
-            ZonedDateTime.ofInstant(Instant.EPOCH, ZoneId.systemDefault())
         private const val NUMBER_OF_DAYS_THE_GREEN_KEYS_ARE_KEPT = 14L
     }
 
@@ -62,36 +58,34 @@ class DatabaseCleanupManagerImpl(
             )
 
             val yellowWarningQuarantine = configuration.yellowWarningQuarantine?.toLong()
-            val thresholdYellowMessageAsRollingStart = if (yellowWarningQuarantine != null) {
-                ZonedDateTime.now()
+            if (yellowWarningQuarantine != null) {
+                val thresholdYellowMessageAsRollingStart = ZonedDateTime.now()
                     .startOfTheUtcDay()
                     // +1 hour buffer time to avoid removing from midnight, since at this moment (12-Jun-2020)
                     // the rolling start interval is set by the framework at midnight (start of the day)
                     .minusHours(yellowWarningQuarantine + 1)
                     .toRollingStartIntervalNumber()
-            } else {
-                UNIX_TIME_START.toRollingStartIntervalNumber()
+
+                temporaryExposureKeysDao.removeSentInfectionMessagesOlderThan(
+                    MessageType.InfectionLevel.Yellow,
+                    thresholdYellowMessageAsRollingStart
+                )
             }
-            temporaryExposureKeysDao.removeSentInfectionMessagesOlderThan(
-                MessageType.InfectionLevel.Yellow,
-                thresholdYellowMessageAsRollingStart
-            )
 
             val redWarningQuarantine = configuration.redWarningQuarantine?.toLong()
-            val thresholdRedMessagesAsRollingStart = if (redWarningQuarantine != null) {
-                ZonedDateTime.now()
+            if (redWarningQuarantine != null) {
+                val thresholdRedMessagesAsRollingStart = ZonedDateTime.now()
                     .startOfTheUtcDay()
                     // +1 hour buffer time to avoid removing from midnight, since at this moment (12-Jun-2020)
                     // the rolling start interval is set by the framework at midnight (start of the day)
                     .minusHours(redWarningQuarantine + 1)
                     .toRollingStartIntervalNumber()
-            } else {
-                UNIX_TIME_START.toRollingStartIntervalNumber()
+
+                temporaryExposureKeysDao.removeSentInfectionMessagesOlderThan(
+                    MessageType.InfectionLevel.Red,
+                    thresholdRedMessagesAsRollingStart
+                )
             }
-            temporaryExposureKeysDao.removeSentInfectionMessagesOlderThan(
-                MessageType.InfectionLevel.Red,
-                thresholdRedMessagesAsRollingStart
-            )
         }
     }
 }
